refactor: moving EditTodoController to its own file
